RUSTSEC-2019-0031: spin is maintained (#424)
We added `yanked = true` to the advisory, however it doesn't seem to be
having the intended effect (the query for unmaintained crates is
probably failing to exclude the yanked advisories)

This is another workaround which makes the `unaffected` requirement
match all versions. Hopefully this means that `spin` will stop being
reported as unmaintained.