Add unmaintained advisory for diesel-adapter
`diesel-adapter` has some dependencies with CVEs open it, and an attempt to update things from a crate consumer shows that (a) there really isn't anyone assigned to maintenance of the crate, e.g. issue fixes are requested from the community rather than having someone able to response, and (b) the source has a broken CI that can't realistically be addressed from a few drive-by patches.