Add advisory for libp2p-tokio-socks5
I was the original author and am still the crate owner (on crates.io),
I do not, however have write access to the source repository because I
no longer work for the company that controls it.

I stopped maintaining this crate in 2021 but did not know about the
RustSec Advisory process then.