Advisory for GHSA-v26r-4c9c-h3j6 (config scopes) in gix-path (#2055)
* Advisory for GHSA-v26r-4c9c-h3j6 (config scopes) in gix-path

* Fix a commit hash intended to be a link to commit info

This worked on GitHub but should not be expected to be a hyperlink
elsehwere. So this makes the rendered text and target explicit.

* Add CVSS metadata

It is present in GHSA-v26r-4c9c-h3j6, I just accidentally left it
out initially.

* Clarify some wording in the advisory

I have already made this change at GHSA-v26r-4c9c-h3j6.