Add global GHSA reference for RUSTSEC-2024-0367 (config scopes) (#2063)
* Add global GHSA reference for RUSTSEC-2024-0367 (config scopes)

This adds a link to the GitHub Advisory Database entry
https://github.com/advisories/GHSA-v26r-4c9c-h3j6 for
RUSTSEC-2024-0367 / CVE-2024-45305 / GHSA-v26r-4c9c-h3j6.

This entry was added to the GitHub Advisory Database since this
RUSTSEC entry was created in #2055 and updated in #2061.

(This also adds a reference to NVD entry, which has a useful
summary and appears as a reference in the global GHSA's reference
section.)

* Linkify bare URLs

The advisory, RUSTSEC-2024-0367, has two bare URLs in it, which
are displayed as links (and, in the repo-level GHSA, also showing
the linked-to lines of code). This surrounds them with `<` and `>`
so that they are rendered as hyperlinks, as they are in the global
GHSA.

(This does not correspond to a revision to the global GHSA because
they are already shown that way there. This change thus brings the
RUSTSEC advisory in line with the others.)