Specify affected functions versions instead of `"*"`
When including information about affected functions in advisory
metadata, I have used `"*"` in a few advisories when the bounds
were no narrower than the bounds for the vulnerability. But
specific are still needed or preferred in that situation.

This replaces `"*"` in those advisories with more specific bounds.
(For #2193, this has already been done in #2195.)

See discussion in:
https://github.com/rustsec/advisory-db/pull/2193#pullrequestreview-2560582694