Import an old Wasmtime security advisory
[Wasmtime] recently got a [request] to have our security advisories
published on the RustSec database as well. We've got a few old
advisories on here but we haven't been keeping up-to-date with later
advisories. In lieu of automatic imports from GitHub to RustSec we
figured we'd in the interim manually fill in some fields.

In this PR I'm filing a security advisory for [GHSA-88xq-w8cq-xfg7]
which is a 3-year-old advisory at this point. This is to serve as an
example for future imports of Wasmtime advisories so my hope is to get
everything ok, copy this behavior for our other advisories that aren't
listed in RustSec, and then applying this process to future advisories.

[Wasmtime]: https://crates.io/crates/wasmtime
[request]: https://github.com/bytecodealliance/wasmtime/issues/10344
[GHSA-88xq-w8cq-xfg7]: https://github.com/bytecodealliance/wasmtime/security/advisories/GHSA-88xq-w8cq-xfg7